commenting

+++ draw_png.cpp
@@ -633,7 +633,7 @@ uint64_t calcImageSize(const int mapChunksX, const int mapChunksZ, const size_t 
 fsub: brightnessAdjustment
 biom parameter not used
 */
-void setPixel(const size_t x, const size_t y, const uint16_t color, const float fsub, const uint8_t biome)
+void setPixel(const size_t x, const size_t y, const uint16_t color, const float fsub, const uint16_t biome)
 {
 
 	// Sets pixels around x,y where A is the anchor
